Fixed WhitelistExtractor crash
ProtoBuf.Net.Core is now skipped by the FIndBlocks method.

ProtoBuf.Net.Core was problematic in that it referenced version 4.5.0.3 of System.ServiceModels.Primitives, which it cannot find.

An alternative way to fix this would be to handle this exception without rethrowing it. I'm not familiar enough with the past development of this project to know which is the better approach.